+++ cmd/gophermart/pkg/models/models.go
@@ -17,28 +17,28 @@ type User struct {
 type Order struct {
 	Number     string   `json:"number"`
 	Status     string   `json:"status"`
-	Accrual    *float64 `json:"accrual,omitempty"`
+	Accrual    *float32 `json:"accrual,omitempty"`
 	UploadedAt string   `json:"uploaded_at"`
 }
 
 type AccrualResponse struct {
 	Order   string   `json:"order"`
 	Status  string   `json:"status"`
-	Accrual *float64 `json:"accrual,omitempty"`
+	Accrual *float32 `json:"accrual,omitempty"`
 }
 
 type UserBalance struct {
-	Current   float64 `json:"current"`
-	Withdrawn float64 `json:"withdrawn"`
+	Current   float32 `json:"current"`
+	Withdrawn float32 `json:"withdrawn"`
 }
 
 type WithdrawRequest struct {
 	Order string  `json:"order"`
-	Sum   float64 `json:"sum"`
+	Sum   float32 `json:"sum"`
 }
 
 type Withdraw struct {
 	Order       string  `json:"order"`
-	Sum         float64 `json:"sum"`
+	Sum         float32 `json:"sum"`
 	ProcessedAt string  `json:"processed_at"`
 }
